+++ Changelog.md
@@ -1,22 +1,46 @@
 # Changelog
 
-**2.6.6**
-- added `SMTP_TLS` configuration parameter
-- redmine: upgrade to v.2.6.6
+**3.2.0**
+- redmine: upgrade to v.3.2.0
 
-**2.6.5-1**
+**3.1.3**
+- redmine: upgrade to v.3.1.3
+
+**3.1.2**
+- redmine: upgrade to v.3.1.2
+
+**3.1.1**
+- renamed `plugins/init` script to `plugins/post-install.sh`
+- added `plugins/pre-install.sh` script to execute commands before plugin installation
+- redmine: upgrade to v.3.1.1
+
+**3.1.0-2**
+- added support for receiving emails via IMAP
+
+**3.1.0**
+- redmine: upgrade to v.3.1.0
+
+**3.0.4**
+- added `SMTP_TLS` configuration parameter
+- redmine: upgrade to v.3.0.4
+
+**3.0.3-1**
 - install: fix typo in `bundle install` command :facepalm:
 - base image update to fix SSL vulnerability
 
-**2.6.5**
-- redmine: upgrade to v.2.6.5
+**3.0.3**
+- redmine: upgrade to v.3.0.3
 
-**2.6.4**
-- redmine: upgrade to v.2.6.4
+**3.0.2**
+- redmine: upgrade to v.3.0.2
 
-**2.6.3**
+**3.0.1**
 - fix: avoid duplicate cron entries for 'Repository.fetch_changesets'
-- redmine: upgrade to v.2.6.3
+- fix: update the path of 'script/rails' script to 'bin/rails' in v.3.0.0
+- redmine: upgrade to v.3.0.1
+
+**3.0.0**
+- redmine: upgrade to v.3.0.0
 
 **2.6.2**
 - update postgresql client to the latest version